pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "colorize-ASCII"
version = "0.1.6"
description = "This project takes in a png file then a text file and outputs a colored ans file"
requires-python = ">=3.0"
dependencies = [
    "Pillow>=10",
    "numpy>=2",
]
authors = [
    { name = "Ian Robinson", email = "irobinson010@gmail.com" },
]
readme = "README.md"
license = {text = "Apache-2.0"}
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
]

[project.scripts]
ascii-magic = "ascii_magic.image_to_ascii:main"
colorize-ascii = "ascii_magic.colorize_ascii:main"
text-to-ascii = "ascii_magic.text_to_ascii:main"

[tool.setuptools]
package-dir = {"" = "src"}
